requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
python-dotenv==1.0.0
python-docx
pydantic==2.5.2
python-dateutil==2.8.2
email-validator==2.1.0.post1
qdrant-client==1.7.0
openai==1.3.7
rich==13.7.0
httpx==0.24.1
fastapi
uvicorn
python-dotenv
pydantic
pymongo
jose
passlib
bcrypt
openai
qdrant-client
requests
email-validator